removed extra file, updated coldstaking tests
